moving the source message callback up a level, so it doesn't get replaced for each smart source
