+++ app/services/sources/properties-managers/smart-browser-source-manager.ts
@@ -1,16 +1,8 @@
 import { PropertiesManager } from './properties-manager';
 import { Inject } from 'services/core/injector';
 import { WebsocketService } from 'services/websocket';
-import { SseService  } from 'services/server-sent-events';
-import * as obs from '../../../../obs-api';
 import { Subscription } from 'rxjs';
 import { VisionService } from 'services/vision';
-import uuid from 'uuid/v4';
-
-interface ISourceMessage {
-  sourceName: string;
-  message: any;
-}
 
 export class SmartBrowserSourceManager extends PropertiesManager {
   @Inject() private websocketService: WebsocketService;
@@ -19,27 +11,6 @@ export class SmartBrowserSourceManager extends PropertiesManager {
   private sseSub!: Subscription;
 
   init() {
-    obs.NodeObs.RegisterSourceMessageCallback(async(evt: ISourceMessage[]) => {
-      console.log("SmartBrowserSourceManager: Received source message", evt);
-      for (const { sourceName, message } of evt) {
-        if (sourceName !== this.obsSource.name) {
-          continue;
-        }
-        const keys = JSON.parse(message).keys;
-        const tree = this.convertDotNotationToTree(keys);
-        const res = await this.visionService.requestState({ query: tree });
-        const payload = JSON.stringify({
-          type: 'state.update',
-          message: res,
-          key: keys?.join(","),
-          event_id: uuid(),
-        });
-        console.log("SmartBrowserSourceManager: Sending message to source", sourceName, payload);
-        this.obsSource.sendMessage({
-          message: payload
-        });
-      }
-    });
     this.socketSub = this.websocketService.socketEvent.subscribe(e => {
       console.log('WS event', e);
 
@@ -54,21 +25,5 @@ export class SmartBrowserSourceManager extends PropertiesManager {
   destroy() {
     this.socketSub?.unsubscribe();
     this.sseSub?.unsubscribe();
-  }
-
-  private convertDotNotationToTree(states: string[] | string): any {
-    const tree: any = {};
-    const stateArray = Array.isArray(states) ? states : [states];
-    stateArray.forEach(state => {
-      const parts = state.split('.');
-      let current = tree;
-      parts.forEach((part, index) => {
-        if (!current[part]) {
-          current[part] = index === parts.length - 1 ? true : {};
-        }
-        current = current[part];
-      });
-    });
-    return tree;
   }
 }
